Small Improvements and fixes

+++ src/main/java/is/hi/hbv501g/hbv1/Controllers/QuestionController.java
@@ -1,18 +1,14 @@
 package is.hi.hbv501g.hbv1.Controllers;
 
 import is.hi.hbv501g.hbv1.Persistence.Entities.Question;
-import is.hi.hbv501g.hbv1.Persistence.Entities.User;
 import is.hi.hbv501g.hbv1.Services.QuestionService;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-
-import java.util.List;
 
 /**
  * Controller for Question.
@@ -25,7 +21,7 @@ import java.util.List;
 public class QuestionController
 {
     // Variables.
-    private QuestionService questionService;
+    private final QuestionService questionService;
 
 
     /**
@@ -43,8 +39,10 @@ public class QuestionController
     /**
      * Create a new Question.
      *
-     * @param question Question object to save.
-     * @return         Redirect.
+     * @param questionnaireID ID of Questionnaire the user was viewing. Redirects back to overview of that Questionnaire.
+     * @param question        Question object to save.
+     * @param result          BindingResult of form.
+     * @return                Redirect.
      */
     @RequestMapping(value = "editQuestionnaire/{questionnaireID}/addQuestion", method = RequestMethod.POST)
     public String saveQuestion(@PathVariable("questionnaireID") Long questionnaireID, Question question, BindingResult result)
@@ -55,7 +53,7 @@ public class QuestionController
         }
 
         // Add Question to database if no errors.
-        questionService.saveQuestion(question);
+        questionService.saveNewQuestion(question);
 
         return "redirect:/editQuestionnaire/" + questionnaireID;
     }
@@ -64,16 +62,29 @@ public class QuestionController
     /**
      * Delete Question.
      *
-     * @param questionID ID of Question to delete.
-     * @return           Redirect.
+     * @param questionnaireID ID of Questionnaire the user was viewing. Redirects back to overview of that Questionnaire.
+     * @param questionID      ID of Question to delete.
+     * @param session         Current HttpSession.
+     * @return                Redirect.
      */
     @RequestMapping(value = "editQuestionnaire/{questionnaireID}/deleteQuestion/{questionID}", method = RequestMethod.GET)
-    public String deleteQuestion(@PathVariable("questionnaireID") Long questionnaireID, @PathVariable("questionID") Long questionID, Model model) {
-        // If Question exists, delete.
+    public String deleteQuestion(@PathVariable("questionnaireID") Long questionnaireID, @PathVariable("questionID") Long questionID, HttpSession session) {
+        // Get Question from database.
         Question exists = questionService.getQuestionById(questionID);
 
-        if (exists != null) {
-            questionService.deleteQuestionById(questionID);
+        if (exists != null)
+        {
+            // If Question has no connected Questionnaires, then delete.
+            if (exists.getQuestionnaires().isEmpty())
+            {
+                questionService.deleteQuestionById(questionID);
+            }
+
+            // If not, then add the QuestionnaireID to the session to display message on editQuestionnaire page.
+            else
+            {
+                session.setAttribute("questionError", exists.getId());
+            }
         }
 
         return "redirect:/editQuestionnaire/" + questionnaireID;
